Move list of valid, invalid recipes, and recipe dependencies
* The abstract ResourceLoader contains the lists and methods now
